Fix a format warning (debug only)
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30869 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/impexp/xp/ie_exp_HTML_util.cpp
@@ -636,7 +636,7 @@ void IE_Exp_HTML_TagWriter::openTag(const std::string& tagName, bool isInline, b
     }
     m_buffer += "<" + tagName;
     
-    UT_DEBUGMSG(("Opened tag: %s Depth: %d\n", tagName.c_str(), m_tagStack.size())); 
+    UT_DEBUGMSG(("Opened tag: %s Depth: %ld\n", tagName.c_str(), (long)m_tagStack.size())); 
 }
 
 void IE_Exp_HTML_TagWriter::addAttribute(const std::string& name, const std::string& value)
